requirements.txt
pymupdf
openai
pydantic
instructor
pandas
numpy
python-dotenv